# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocSmith — AI-powered Universal File-to-DOCX Converter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversionStatus =
  (typeof ConversionStatus)[keyof typeof ConversionStatus];

export const ConversionStatus = {
  pending: "pending",
  processing: "processing",
  done: "done",
  error: "error",
} as const;

export interface Conversion {
  id: number;
  originalFilename: string;
  originalExtension: string;
  fileSize: number;
  status: ConversionStatus;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  downloadPath: string | null;
  /** @nullable */
  processingMode: string | null;
  useAi: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ExtensionCount {
  extension: string;
  count: number;
}

export interface ConversionStats {
  total: number;
  done: number;
  error: number;
  pending: number;
  byExtension: ExtensionCount[];
}

export interface ErrorResponse {
  error: string;
}