# lib/api-zod/src/generated/types/conversion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocSmith — AI-powered Universal File-to-DOCX Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionStatus } from "./conversionStatus";

export interface Conversion {
  id: number;
  originalFilename: string;
  originalExtension: string;
  fileSize: number;
  status: ConversionStatus;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  downloadPath: string | null;
  /** @nullable */
  processingMode: string | null;
  useAi: boolean;
  createdAt: Date;
  updatedAt: Date;
}