# lib/api-zod/src/generated/types/conversionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DocSmith — AI-powered Universal File-to-DOCX Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtensionCount } from "./extensionCount";

export interface ConversionStats {
  total: number;
  done: number;
  error: number;
  pending: number;
  byExtension: ExtensionCount[];
}